* Serialization/deserialization classes are moved to the same place
