+++ Replicated_Storage_Service/src/testing/weloveclouds/kvstore/serialization/ServerConnectionInfosSetTest.java
@@ -11,10 +11,10 @@ import junit.framework.Assert;
 import junit.framework.TestCase;
 import weloveclouds.commons.kvstore.deserialization.exceptions.DeserializationException;
 import weloveclouds.commons.kvstore.deserialization.helper.IDeserializer;
+import weloveclouds.commons.kvstore.deserialization.helper.ServerConnectionInfosSetDeserializer;
 import weloveclouds.commons.kvstore.serialization.helper.ISerializer;
+import weloveclouds.commons.kvstore.serialization.helper.ServerConnectionInfosSetSerializer;
 import weloveclouds.communication.models.ServerConnectionInfo;
-import weloveclouds.kvstore.deserialization.helper.ServerConnectionInfosSetDeserializer;
-import weloveclouds.kvstore.serialization.helper.ServerConnectionInfosSetSerializer;
 import weloveclouds.server.utils.SetToStringUtility;
 
 /**
