+++ Replicated_Storage_Service/src/weloveclouds/commons/kvstore/deserialization/helper/MovableStorageUnitsSetDeserializer.java
@@ -1,4 +1,4 @@
-package weloveclouds.kvstore.deserialization.helper;
+package weloveclouds.commons.kvstore.deserialization.helper;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -6,9 +6,7 @@ import java.util.Set;
 import org.apache.log4j.Logger;
 
 import weloveclouds.commons.kvstore.deserialization.exceptions.DeserializationException;
-import weloveclouds.commons.kvstore.deserialization.helper.IDeserializer;
-import weloveclouds.commons.kvstore.deserialization.helper.MovableStorageUnitDeserializer;
-import weloveclouds.kvstore.serialization.helper.MovableStorageUnitsSetSerializer;
+import weloveclouds.commons.kvstore.serialization.helper.MovableStorageUnitsSetSerializer;
 import weloveclouds.server.store.models.MovableStorageUnit;
 
 /**
